# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Concert API",
        "description": "API for managing concerts",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/concerts/{id}": {
            "get": {
                "tags": [
                    "Concerts"
                ],
                "summary": "Get a specific concert",
                "operationId": "7a300c3da742086037781e7f07a39725",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Concert ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concert details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "year": {
                                            "type": "integer"
                                        },
                                        "type": {
                                            "type": "string"
                                        },
                                        "source": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Concert not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Concerts"
                ],
                "summary": "Update a concert",
                "operationId": "db80212e0a8f5b07e59773988e4fdd2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Concert ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "year": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "concert",
                                            "festival",
                                            "dj set",
                                            "club show",
                                            "theater show"
                                        ]
                                    },
                                    "source": {
                                        "type": "string",
                                        "enum": [
                                            "manual",
                                            "api"
                                        ]
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending_approval",
                                            "verified",
                                            "rejected"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Concert updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "year": {
                                            "type": "integer"
                                        },
                                        "type": {
                                            "type": "string"
                                        },
                                        "source": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Concert not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Concerts"
                ],
                "summary": "Delete a concert",
                "operationId": "7d8b63d823379ebfa22ae8310590d030",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Concert ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Concert deleted successfully"
                    },
                    "404": {
                        "description": "Concert not found"
                    }
                }
            }
        },
...
    },
...
}